Use new and improved math for retopology offset
The clip distances no longer need to be factored in on the C++ side, so that simplifies how bias is calculated.
I've added a new function to the common_view_lib, which uses math I found in a neat PDF for perspective, and math I came up with through a little trial and error for orthographic.
The shaders that use retopology bias now have an extra step to get the view position, but that was being calculated to get the homogenous position anyway so it's just an extra variable.
The offset is done at the start of the vertex shader now instead of the end; I think it makes sense because the offset it does is usually much larger than anything that comes after.