Implement retopology overlay requests
The bias property is now called offset, not just in the UI but everywhere, because I like consistency.
get_homogenous_z_offset no longer negates its output, and has some more documentation.
The offset property is now of distance type, because it is a viewspace offset.
To make this more clear the soft max is increased to 10.
Because there's no point in arbitrarily limiting the values users can input, the hard max is FLT_MAX now.
I kept step the same, because it good the way it is.