UI: add parenting (Keep Transform Without Inverse) to menu
Since db6287873cd2, we added parenting (Keep Transform Without Inverse)
to the parent_set operator and exposed this in the UI in the invoke menu
of that operator.

This however never made it to the `Object` > `Parent` menu.

Since this is quite useful (and otherwise a bit hidden in the Adjust
Last Operation panel) and parent inverse can cause confusion in many
cases (see e.g. #105276 or #103969), now add this as an explicit entry
in the `Object` > `Parent` menu as well.

NOTE: arguably we could unify the entries in the invoke menu and the
`Object` > `Parent` menu as well since these are both different:
- invoke calls this
-- Set Parent to > Object (Without Inverse)
-- Set Parent to > Object (Keep Transform Without Inverse)
- `Object` > `Parent` menu
-- Make Parent without Inverse
-- Make Parent without Inverse (Keep Transform)