Fix: Undo not working properly with Mesh Filter
The previous behaviour was that if you applied a mesh filter and
modified its strength for example through the redo last operator, it
would additively add the new update on top of the existing one. It would
also add an undo step.

This was due to calling `SCULPT_undo_push_end_ex(ob, true)` on both
operator confirmation and cancellation, where it was only needed on
cancellation.