UI: Add scroll to sidebar tabs
When the height of the editor couldn’t fit the sidebar tabs, they would shrink
to a minimum size that made it impossible to read the tab labels.

This pull request matches the behaviour with the Properties Editor navigation
bar, by introducing the following improvements:
* Avoid truncating tab labels.
* Allow scrolling when tabs don’t fit.

Behaviour is similar to how scrolling works in the Properties Editor navigation
bar, supporting mouse wheel up/down and MMB, and switching tabs with
`Ctrl+Wheel Up/Down`.