UI: Make it possible to add shortcuts to UI operators.
Note: This patch originates from investigation of #105371. I have noticed, thatthis has been intentionally disabled, but it was long time ago without any comment why and `PAINT_OT` already violates this. So consider it as if I was proposing this blindly.
Perhaps it make sense to exclude specific UI operators if some are problematic?