Bone relation lines: draw between axis points
Change how bone relation lines are drawn. Instead of drawing from the head
of the child to the tail of the parent, use the "axis points" as start/end
point of the lines. This is the point used to draw the bone axes, and
thus can be modified via the "Axes Position" setting.