Image: Add OIIO support APIs
This adds a new set of APIs supporting the loading and saving of image
formats through OIIO. It makes use of the recent IOProxy work in OIIO
to align with the existing Blender image loading/saving machinery.

The support code here has been prototyped to work with ~7 of our image
formats so far[1]. It includes centralized handling of `IB_test`,
`IB_mem`, and `IB_metadata` flags, which the existing code did not
handle consistently or at all depending on the format.

The PSD format (`format_psd.cc`) is included here since the prior
location of the code has been restructured away. It serves as an example
of how the loading code typically flows for all the other formats.[2]

----
This continues the work for #101413

[1] The BMP,DDS,HDR,PNG,TIFF,and TARGA formats will be sent for review after this
[2] Incidentally this will address #70495 for "free"