Fix: No image GPU texture update for different requests
The image_get_gpu_texture function checks if the requested properties
are different from the last requested ones, yet it does nothing to
update the texture if it is indeed the case. This patch marks the image
for an update in that case. While marking an update during a request
might seem out of place, that full update marking is immediately handled
by freeing the no longer valid cached GPU texture in
image_gpu_texture_try_partial_update().