Cycles: Adjust resolution divider to achieve a more usable viewport
This patch changes the maximum viewport resolution divider for Cycles
to help ensure users get a more useable viewport.

This is done by changing the maximum viewport resolution divider
to a divider that aims to have the largest axis of the viewport
roughly equal to 128 pixels.

Depending on the circumstances, this change can result in a few
noticeable differences:
 - Users with slow hardware and a large pixel_size, or slow hardware
 and a low resolution screen, may observe a higher resolution viewport
 during navigation, making the scene more readable. However this comes
 at the cost of reduced performance.

 - Users with slow hardware and a low pixel_size may observe a
 lower resolution viewport during navigation, providing a more
 responsive viewport during navigation.

---

Along with that, how Cycles iterates through resolution dividers
is changed to promote quick transitions between resolution dividers.
Meaning users don't need to wait through as many iterations to get
from a low navigation resolution to a 1:1 viewport resolution.