Replace the 0-1 slider with an armature flag
The choice between drawing relation lines from the parent head or tail is
now a bitflag on the armature. There were plenty of unused flags left, so
it now doesn't take up any space any more.

The flag is wrapped in an RNA enum, so that there are distinct choices
'Head' and 'Tail'. This is preferred over a single checkbox.