I18n: add per-label translation contexts for nodes
In order to properly translate UI messages, they sometimes need to be
disambiguated using translation contexts. Until now, node sockets
using custom labels had no way to specify contexts and a collision
occurred in at least one known instance.

This commit adds a way to declare contexts for each socket with
labels, by adding a `label_translation_context` field to the
bNodeSocket struct, as well as a `translation_context` argument to the
`node_sock_label()` function.

If no context is specified, the default null context is used.

The use of a context will enable us to fix part of #105113 in the next
commit.