Vulkan: Add initial VkPixelBuffer.
VK Pixel buffer is used by external render engines to allocate
buffers using the same GPU backend that Blender runs in.

In a later stage we should test what exact binding flags are
needed. I expect that it should be able to use as a transfer
buffer to copy the pixels over to a texture using transfer
commands.