Python API: Expose background drawing as argument for gpu.types.GPUOffScreen.draw_view3d()
Currently, when using the python api for offscreen drawing, the default background will always be rendered into the GPUOffScreen's framebuffer, rendering the alpha channel essentially useless and making it difficult to separate objects from the background.

This patch allows offscreen drawing of a 3d view with transparent background by exposing an optional parameter to the python api, enabling, for example, compositing the result over another image.

The new parameter to draw_view3d() is optional, with the default value matching the previous behavior, so this change is fully backwards-compatible.