Enable the balanced length constraint solver.
This gets a preconditioner to drastically speed up convergence
in case curves get stretched. In that case the overall distance
to the root can be used to get close to the solution.

Without this preconditioning a few iterations (currently 5) are
not enough to get close to constance segment lengths, and with
every new stroke the curves stretch a bit more. Preserving
goal length as an attribute in future should also help with this
problem of drift.