Mesh: Cache vertex to face topology map
This commit uses the `SharedCache` concept to keep a cached map from
mesh vertices to polygons available on a mesh as necessary. The map
is split into two caches internally, since the same offsets can be
reused for a potential vertex to face corner map in the future
(though that may be a bit preemptive).

Such a cache (or similar) is often used in topology-changing operations,
and in the future it can be updated as necessary. For now it's simply
calculated as necessary though, since it may take time to validate the
choice of cached map. For example, it could make sense to cache
a vertex to corner and a corner to face map instead.