Restore active/default color attribute using index
Find the indices for the active and default color attributes, remove the chosen attribute, find the color attributes that go with those indices now, and make them active/default.
This code is not exactly elegant but it does work.
A utility function would be nice, instead of copy pasting the same code several times.
But it's late and this is good enough for a proof of concept.