Realtime Compositor: Implement Fog Glow Glare node
This patch implements the Fog Glow option in the Glare node. The
implementation does not match the existing implementation in the CPU
compositor, because it is computationally infeasible for the realtime
compositor. Instead, this implementation is similar to how Bloom is
implemented in EEVEE.

Since the Glare node has its own threshold and mix steps, the blit and
resolve steps were omitted from the EEVEE implementation, leaving just
the upsampling and downsampling chains. So one should not expect this to
be identical to EEVEE's result.

This is just a temporary solution as requested by users until we either
implement an accurate implementation using FFT, provide a separate
option for EEVEE bloom, or roll out a different solution altogether.