Cycles: compute unique light sets for light linking
We plan to build specialized light trees for every unique set of light
emitters. This requires knowing unique combinations of lights used by
receiver. However for easier library linking and USD compatibility the
relation is stored in the opposite direction.

This modifies the depsgraph and Cycles code to compute and use a unique
bit for every combination of lights.