UI: Flip Side Regions
The current UI allows users to reposition certain regions like the header,
but there are no easy way to customize the side regions to suit specific
workflows. For instance, right-handed users with graphical tablets in
sculpting mode often have to cross the screen with their right hand to
access the required tool.

To address this issue, it's introduced a menu and operator that allows
users to easily swap the position of the side regions. This new feature
provides users additional options to fit the interface to their unique
workflows.