Geometry: Skip recomputing bounds after translation
Translating meshes, point clouds, and curves and recomputing the bounds
is equivalent to just translating the bounds. Now that mesh primitive
nodes calculate their bounds in constant time, it's worth updating the
bounds eagerly when translating a geometry since doing so should be
very cheap and might save a more significant amount of time if they're
needed later.