Fix: Compiling Shader Builder.
IMath is required for building with vulkan. This patch adds openexr
dependencies to shader_builder for now. Imath is a dependency of
openexr.

Currently there is a difference between windows and other OS's in
which variable the library is stored. Might be because imath isn't
a direct dependecy. As this is a compilation error I did the quickest
solution and check with platform module what would be the best solution.