Font: reorganize Special Characters menu with submenus
Many new characters are introduced, split into 10 submenus: Math,
Fraction, Subscript, Superscript, Quotation Mark, Dash, Punctuation,
Currency, Intellectual Property.

The choice of characters included is of course arbitrary, but not any
more than before. It is based on what I believe may be most commonly
useful for users, or at least users using Western Scripts.

They are all characters which are hard to input with an ordinary
keyboard layout (QWERTY, AZERTY, QWERTZ, etc.) as opposed to
nerd-friendly, Dvorak-type layouts.

Since this is not so common an operation, I believe it warrants adding
a new level of menus so as not to have dozens of characters in one
menu.

- Many common math symbols were added, as well as common fractions.
- Superscript 1, 2, 3 were extended to 0, and completed by subscripts.
- Copyright and Registered Trademark were completed by TM and Sound
  Recording Copyright.
- "Double <<" and "Double >>" were moved to a "Quotation" menu and
  renamed to Left and Right Guillemet, respectively, as they are
  unrelated to the less than and greater than signs. These were
  completed by other characters useful for quoting in various languages.
- The Promillage was moved to Math symbols and renamed "Per Mille" as
  it seems to actually be Dutch (?).
- Circle was removed, as the character was invalid and had apparently
  been for many years. It is not obvious what circle character it was
  intended to add, and this is not such a common sign anyway.
- The Dutch Florin, British Pound and Japanese Yen were stripped of
  their countries, because the symbols may be used for other
  currencies. In fact the Florin was renamed to "Currency Sign" as
  that is what it is. They were moved to their own menu, along with
  new characters Cent, Euro, and Dollar--although this last one is easy to
  input with any ordinary keyboard.
- Eszett ("German S") was removed, based on the assumption that people
  most likely to use it are German speakers and probably use or can
  switch to a layout enabling them to input it.
- "Spanish question mark" and "Spanish exclamation mark" were moved to
  a "Punctuation" menu, along with many new characters.
- A new Dash menu was added.

Each menu entry now shows the name of the character, followed by the
char itself, to make it easier to choose visually.